Add timeout for account authenticator connection
When the system server is bound to the application's authenticator, if the application does not call the onResult method, the Service connection between the system server and the application will not be disconnected, and the application process will maintain high priority for a long time and cannot be recycled. So timeout disconnection is necessary.

Change-Id: Id1726d4f4b1e47701efde56de2d25ecf70b28ec1
Signed-off-by: ryzenforce990 <ryzenforce990@gmail.com>